Began refactor to seperate api calls.
